Pour résolution de bug
