--- src/HarasBundle/Resources/views/template_footer.html.twig
@@ -1,36 +1,0 @@
-<footer id="_template_footer">
-    <div class="footer-wrapper">
-        <div class="footer-left">
-            <p class="p-titre">Qui sommes nous ?<p>
-            <p> Le Haras de la Métamorphose</p>
-            <p> est un lieu sauvegardé au sein</p>
-            <p> de la forêt de Compiègne où vous</p>
-            <p> pourrez retrouver toutes vos</p>
-            <p> activités chevalines. </p>
-
-        </div>
-        <div class="footer-center">
-            <p class="p-titre">Nos coordonnées</p>
-            <p>Adresse <br />
-            60 000 Compiègne</p>
-            <p>01 46 74 65 07</p>
-            <p class="p-titre"><a id="avis" href="/somwhere.htm">Vos avis</a></p>
-
-
-        </div>
-        <div class="footer-right">
-            <p class="p-titre">Newsletter</p>
-            <form method="POST" action="add_email.php">
-
-                <input required type="email" name="email" placeholder="Abonnez-vous" />
-                <button class="button button1" type="submit">Abonnez-vous</button>
-
-            </form>
-
-        </div>
-    </div>
-    <div id="footer-bottom">
-        <a id="copyright" href="">© 2016 Haras de la Métamorphose</a>
-        <!--				<a id="mentions" href="">Mentions légales</a> -->
-    </div>
-</footer>	<!--  Ferme l'id #_index_footer-->
