--- src/HarasBundle/Resources/views/template_remplissage.html.twig
@@ -1,67 +1,0 @@
-{% extends '@Haras/base.html.twig' %}
-
-{% block materialize %}
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.7/css/materialize.min.css">
-{% endblock materialize %}
-
-{% block body %}
-<body id="template">
-    <header id="_template_header">
-        {% include "HarasBundle:Index:index_header.html.twig" %}
-    </header>
-
-    <div class="article">
-        <div>
-            <h1 class="title">Découverte de la forêt</h1>
-        </div>
-        <div class="bloc1">
-            <table>
-                <tr>
-                    <td>
-                        <h2>Sous-titre 1</h2>
-                        <p id="texte-gauche">
-                            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Commodi culpa cumque esse magnam minima officia rem rerum similique, sunt voluptas! Ab adipisci deleniti dolorem error esse illum iste nostrum qui.         Lorem ipsum dolor sit amet, consectetur adipisicing elit. Nobis, veniam, eaque iusto nostrum excepturi deserunt? Vitae, neque, alias? Ea alias tenetur explicabo assumenda eos, doloremque modi aspernatur odio est distinctio.
-                        </p>
-                    </td>
-                    <td>
-                        <img id="image-droite" src="https://tse3.mm.bing.net/th?id=OIP.M647c41abdcea5ffe31e9ab3ec3621778H0&pid=Api" alt="image_template"/>
-                    </td>
-                </tr>
-            </table>
-        </div>
-        <div class="bloc1">
-            <table>
-                <tr>
-                    <td>
-                        <h2>Sous-titre 1</h2>
-                        <p id="texte-gauche">
-                            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Nobis, veniam, eaque iusto nostrum excepturi deserunt? Vitae, neque, alias? Ea alias tenetur explicabo assumenda eos, doloremque modi aspernatur odio est distinctio.
-                        </p>
-                    </td>
-                    <td>
-                        <img id="image-droite" src="https://tse3.mm.bing.net/th?id=OIP.M647c41abdcea5ffe31e9ab3ec3621778H0&pid=Api" alt="image_template"/>
-                    </td>
-                </tr>
-            </table>
-        </div>
-        <div class="bloc2">
-            <table>
-                <tr>
-                    <td>
-                        <h2>Sous-titre 2</h2>
-                        <img id="image-gauche" src="https://tse3.mm.bing.net/th?id=OIP.M647c41abdcea5ffe31e9ab3ec3621778H0&pid=Api" alt="image_template"/>
-                    </td>
-                    <td>
-                        <p id="texte-droite">
-                            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Nobis, veniam, eaque iusto nostrum excepturi deserunt? Vitae, neque, alias? Ea alias tenetur explicabo assumenda eos, doloremque modi aspernatur odio est distinctio.
-                        </p>
-                    </td>
-                </tr>
-            </table>
-        </div>
-    </div>
-
-    {%  include 'HarasBundle::template_footer.html.twig' %}
-
-
-{% endblock body %}
